Change import statements due to refactor
